+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 2.7.1
+
+- No significant changes since `2.7.0`.
+
 ## 2.7.0
 
 - Remove no-alloc support because it removed `ByteSize::display()` when default features were disabled.